Friendly refactoring (Note: includes nitpicking and removal of Yaml support. YMMV.)
