+++ django_hstore/forms.py
@@ -1,49 +1,47 @@
-try:
-    import yaml
-    def _to_python(value):
-        return yaml.load(value)
-    def _to_text(value):
-        return yaml.dump(value, default_flow_style=False)
-except ImportError:
-    try:
-        import json
-    except ImportError:
-        from django.utils import simplejson as json
-    def _to_python(value):
-        return json.loads(value)
-    def _to_text(value):
-        return json.dumps(value, sort_keys=True, indent=2)
-
-from django.contrib.admin.widgets import AdminTextareaWidget
 from django.forms import Field
+from django.utils import simplejson as json
+from django.contrib.admin.widgets import AdminTextareaWidget
 
 from django_hstore import util
 
-class DictionaryFieldWidget(AdminTextareaWidget):
+class JsonMixin(object):
+
+    def to_python(value):
+        return json.loads(value)
+
     def render(self, name, value, attrs=None):
-        return super(DictionaryFieldWidget, self).render(name, _to_text(value), attrs)
+        value = json.dumps(value, sort_keys=True, indent=2)
+        return super(JsonMixin, self).render(name, value, attrs)
 
-class DictionaryField(Field):
-    """A dictionary form field."""
 
+class DictionaryFieldWidget(JsonMixin, AdminTextareaWidget):
+    pass
+
+
+class ReferencesFieldWidget(JsonMixin, AdminTextareaWidget):
+
+    def render(self, name, value, attrs=None):
+        value = util.serialize_references(value)
+        return super(ReferencesFieldWidget, self).render(name, value, attrs)
+
+
+class DictionaryField(JsonMixin, Field):
+    """
+    A dictionary form field.
+    """
     def __init__(self, **params):
         params['widget'] = DictionaryFieldWidget
         super(DictionaryField, self).__init__(**params)
 
-    def to_python(self, value):
-        return _to_python(value)
 
-class ReferencesFieldWidget(AdminTextareaWidget):
-    def render(self, name, value, attrs=None):
-        value = util.serialize_references(value)
-        return super(ReferencesFieldWidget, self).render(name, _to_text(value), attrs)
-
-class ReferencesField(Field):
-    """A references form field."""
-
+class ReferencesField(JsonMixin, Field):
+    """
+    A references form field.
+    """
     def __init__(self, **params):
         params['widget'] = ReferencesFieldWidget
         super(ReferencesField, self).__init__(**params)
 
     def to_python(self, value):
-        return util.unserialize_references(_to_python(value))
+        value = super(ReferencesField, self).to_python(value)
+        return util.unserialize_references(value)
