+++ django_hstore/query.py
@@ -1,18 +1,12 @@
 from django import VERSION
 from django.db import transaction
-from django.db.models.manager import Manager
 from django.db.models.query import QuerySet
 from django.db.models.sql.constants import SINGLE
 from django.db.models.sql.datastructures import EmptyResultSet
 from django.db.models.sql.query import Query
 from django.db.models.sql.subqueries import UpdateQuery
-from django.db.models.sql.where import EmptyShortCircuit, WhereNode
+from django.db.models.sql.where import EmptyShortCircuit, WhereNode, QueryWrapper
 
-class literal_clause(object):
-    def __init__(self, sql, params):
-        self.clause = (sql, params)
-    def as_sql(self, qn, connection):
-        return self.clause
 
 def select_query(method):
     def selector(self, *args, **params):
@@ -21,6 +15,7 @@ def select_query(method):
         query.clear_select_fields()
         return method(self, query, *args, **params)
     return selector
+
 
 def update_query(method):
     def updater(self, *args, **params):
@@ -47,6 +42,7 @@ def update_query(method):
 
     updater.alters_data = True
     return updater
+
 
 class HStoreWhereNode(WhereNode):
     def make_atom(self, child, qn, connection):
@@ -77,12 +73,13 @@ class HStoreWhereNode(WhereNode):
                     raise ValueError('invalid value')
             else:
                 raise TypeError('invalid lookup type')
-        else:
-            return super(HStoreWhereNode, self).make_atom(child, qn, connection)
+        return super(HStoreWhereNode, self).make_atom(child, qn, connection)
+
 
 class HStoreQuery(Query):
     def __init__(self, model):
         super(HStoreQuery, self).__init__(model, HStoreWhereNode)
+
 
 class HStoreQuerySet(QuerySet):
     def __init__(self, model=None, query=None, using=None):
@@ -91,48 +88,52 @@ class HStoreQuerySet(QuerySet):
 
     @select_query
     def hkeys(self, query, attr):
-        """Enumerates the keys in the specified hstore."""
-
+        """
+        Enumerates the keys in the specified hstore.
+        """
         query.add_extra({'_': 'akeys("%s")' % attr}, None, None, None, None, None)
         result = query.get_compiler(self.db).execute_sql(SINGLE)
         return (result[0] if result else [])
 
     @select_query
     def hpeek(self, query, attr, key):
-        """Peeks at a value of the specified key."""
-
+        """
+        Peeks at a value of the specified key.
+        """
         query.add_extra({'_': '%s -> %%s' % attr}, [key], None, None, None, None)
         result = query.get_compiler(self.db).execute_sql(SINGLE)
         if result and result[0]:
             field = self.model._meta.get_field_by_name(attr)[0]
             return field._value_to_python(result[0])
 
-    @update_query
-    def hremove(self, query, attr, keys):
-        """Removes the specified keys in the specified hstore."""
-
-        value = literal_clause('delete("%s", %%s)' % attr, [keys])
-        field, model, direct, m2m = self.model._meta.get_field_by_name(attr)
-        query.add_update_fields([(field, None, value)])
-        return query
-
     @select_query
     def hslice(self, query, attr, keys):
-        """Slices the specified key/value pairs."""
-
+        """
+        Slices the specified key/value pairs.
+        """
         query.add_extra({'_': 'slice("%s", %%s)' % attr}, [keys], None, None, None, None)
         result = query.get_compiler(self.db).execute_sql(SINGLE)
         if result and result[0]:
             field = self.model._meta.get_field_by_name(attr)[0]
             return dict((key, field._value_to_python(value)) for key, value in result[0].iteritems())
-        else:
-            return {}
+        return {}
+
+    @update_query
+    def hremove(self, query, attr, keys):
+        """
+        Removes the specified keys in the specified hstore.
+        """
+        value = QueryWrapper('delete("%s", %%s)' % attr, [keys])
+        field, model, direct, m2m = self.model._meta.get_field_by_name(attr)
+        query.add_update_fields([(field, None, value)])
+        return query
 
     @update_query
     def hupdate(self, query, attr, updates):
-        """Updates the specified hstore."""
-
-        value = literal_clause('"%s" || %%s' % attr, [updates])
+        """
+        Updates the specified hstore.
+        """
+        value = QueryWrapper('"%s" || %%s' % attr, [updates])
         field, model, direct, m2m = self.model._meta.get_field_by_name(attr)
         query.add_update_fields([(field, None, value)])
         return query
