+++ tests/app/models.py
@@ -1,4 +1,3 @@
-
 from django.db import models
 from django_hstore import hstore
 
@@ -11,7 +10,8 @@ class Ref(models.Model):
 class DataBag(models.Model):
     name = models.CharField(max_length=32)
     data = hstore.DictionaryField(db_index=True)
-    objects = hstore.Manager()
+
+    objects = hstore.HStoreManager()
 
     def __unicode__(self):
         return self.name
@@ -19,7 +19,8 @@ class DataBag(models.Model):
 class RefsBag(models.Model):
     name = models.CharField(max_length=32)
     refs = hstore.ReferencesField(db_index=True)
-    objects = hstore.Manager()
+
+    objects = hstore.HStoreManager()
 
     def __unicode__(self):
         return self.name
